requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scikit-learn>=1.2.0
tqdm>=4.65.0
python-dotenv>=1.0.0
nlpaug>=1.1.10
sentencepiece>=0.1.99
protobuf>=3.20.0
openai==1.3.0
spacy==3.7.2
datasets==2.15.0
evaluate==0.4.1
seqeval==1.2.2
sacremoses==0.1.1
wandb==0.16.0
accelerate==0.25.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
anthropic>=0.7.0
keybert>=0.8.3
presidio-analyzer>=2.2.33
presidio-anonymizer>=2.2.33
pytest>=8.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pandas>=2.2.0 